Extract reset_system_identifier and freeze_master_data to Greenplum-specific file.
