+++ src/bin/pg_upgrade/Makefile
@@ -14,7 +14,8 @@ OBJS = check.o controldata.o dump.o exec.o file.o function.o info.o \
 # Source files specific to Greenplum
 OBJS += greenplum/aotable.o greenplum/gpdb4_heap_convert.o greenplum/version_gp.o \
 		greenplum/check_gp.o greenplum/file_gp.o greenplum/reporting.o \
-		greenplum/aomd_filehandler.o greenplum/option_gp.o
+		greenplum/aomd_filehandler.o greenplum/option_gp.o \
+		greenplum/controldata_gp.o
 
 override CPPFLAGS := -DDLSUFFIX=\"$(DLSUFFIX)\" -I$(srcdir) -I$(libpq_srcdir) $(CPPFLAGS)
 LDFLAGS += -L$(top_builddir)/src/fe_utils -lpgfeutils -lpq
